requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
idna==2.10
imgkit==1.0.2
multidict==5.1.0
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wkhtmltopdf==0.2
yarl==1.6.3
